+++ src/test/java/mocacong/server/service/FavoriteServiceTest.java
@@ -13,12 +13,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -92,35 +87,5 @@ class FavoriteServiceTest {
 
         assertThrows(NotFoundFavoriteException.class,
                 () -> favoriteService.delete(member.getEmail(), cafe.getMapId()));
-    }
-
-    @Test
-    @DisplayName("회원이 한 카페를 동시에 여러 번 즐겨찾기 등록 시도해도 한 번만 등록된다")
-    void saveFavoriteWithConcurrent() throws InterruptedException {
-        Member member = new Member("kth990303@naver.com", "encodePassword", "케이", "010-1234-5678");
-        memberRepository.save(member);
-        Cafe cafe = new Cafe("2143154352323", "케이카페");
-        cafeRepository.save(cafe);
-        ExecutorService executorService = Executors.newFixedThreadPool(3);
-        CountDownLatch latch = new CountDownLatch(3);
-        List<Throwable> exceptions = Collections.synchronizedList(new ArrayList<>());
-
-        for (int i = 0; i < 3; i++) {
-            executorService.execute(() -> {
-                try {
-                    favoriteService.save(member.getEmail(), cafe.getMapId());
-                } catch (AlreadyExistsFavorite e) {
-                    exceptions.add(e); // 중복 예외를 리스트에 추가
-                }
-                latch.countDown();
-            });
-        }
-        latch.await();
-
-        List<Favorite> favorites = favoriteRepository.findAll();
-        assertAll(
-                () -> assertThat(exceptions.isEmpty()).isFalse(),
-                () -> assertThat(favorites).hasSize(1)
-        );
     }
 }
