+++ src/test/java/mocacong/server/service/MemberServiceTest.java
@@ -23,12 +23,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 import static java.lang.Integer.parseInt;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -505,33 +500,6 @@ class MemberServiceTest {
                 () -> assertThat(actual.getEmail()).isEqualTo(email),
                 () -> assertThat(actual.getNickname()).isEqualTo(nickname),
                 () -> assertThat(actual.getPhone()).isEqualTo(phone)
-        );
-    }
-
-    @Test
-    @DisplayName("회원이 동시에 여러 번 가입 시도해도 한 번만 가입된다")
-    void signUpWithConcurrent() throws InterruptedException {
-        MemberSignUpRequest request = new MemberSignUpRequest("kth990303@naver.com", "a1b2c3d4", "케이", "010-1234-5678");
-        ExecutorService executorService = Executors.newFixedThreadPool(3);
-        CountDownLatch latch = new CountDownLatch(3);
-        List<Throwable> exceptions = Collections.synchronizedList(new ArrayList<>());
-
-        for (int i = 0; i < 3; i++) {
-            executorService.execute(() -> {
-                try {
-                    memberService.signUp(request);
-                } catch (DuplicateMemberException e) {
-                    exceptions.add(e); // 중복 예외를 리스트에 추가
-                }
-                latch.countDown();
-            });
-        }
-        latch.await();
-
-        List<Member> actual = memberRepository.findAll();
-        assertAll(
-                () -> assertThat(exceptions.isEmpty()).isFalse(),
-                () -> assertThat(actual).hasSize(1)
         );
     }
 }
